backend/requirements.txt
asgiref==3.7.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
contourpy==1.1.0
cryptography==41.0.1
cycler==0.11.0
defusedxml==0.7.1
Django==4.2.2
django-allauth==0.54.0
django-cors-headers==4.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.6
fonttools==4.40.0
idna==3.4
importlib-resources==5.12.0
inflection==0.5.1
joblib==1.3.0
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.25.0
oauthlib==3.2.2
packaging==23.1
Pillow==9.5.0
pycparser==2.21
PyJWT==2.7.0
pyparsing==3.1.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
scikit-learn==1.2.2
scipy==1.11.1
six==1.16.0
sqlparse==0.4.4
threadpoolctl==3.1.0
typing_extensions==4.7.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.3
zipp==3.15.0
